frontend: move whole-page components into subdir
I'm not 100% on diving the components up this way... or what to call the
directory, but maybe it'll help.
