+++ frontend/src/navbar.tsx
@@ -1,4 +1,4 @@
-import { usePulls, useSetFilter } from './pulls-context';
+import { usePulls, useSetFilter } from './pulldasher/pulls-context';
 import { Pull } from './pull';
 import { Center, Flex, Box, BoxProps, Input } from "@chakra-ui/react"
